Complete streaming of conversion, line-by-line, end-to-end.
