+++ rockstarpy/__main__.py
@@ -1,4 +1,4 @@
-from rockstarpy import command_line
+from .command_line import command_line
 
 if __name__ == '__main__':
     command_line()
